Only add shipment as an order item if present
If there is no shipment information, we currently generate a bogus order item, i.e. things like the `id` value are missing from the item. This causes certain payment processing services to refuse the order. In order to avoid that, we must only add an order item for shipment information, if it is actually present.